+++ jedissource/src/main/java/rahla.jedissource/JedisSource.java
@@ -1,4 +1,4 @@
-package rahla.api;
+package rahla.jedissource;
 
 import redis.clients.jedis.Jedis;